+++ CHANGES.md
@@ -10,6 +10,8 @@
 
 * Fix issue where **Supvisors** may be blocked in the `DEPLOYMENT` phase due to late process events.
 
+* Add a new `start_any_process` XML-RPC that starts one process whose namespec matches the regular expression.
+
 * Add a `wait` parameter to the `update_numprocs` XML-RPC.
 
 * Add the principle of **Supvisors** modes to the output of the XML-RPCs `get_supvisors_state` and `get_instance_info`.